Move hardware modules into hal/, separating Pi-specific from app code
dial.py, positional_encoders.py, buttons.py, rgb_led.py, display.py,
and audio_async.py lived at the top level of the radioglobe package,
mixed in with pure app/data modules (database.py, navigation.py,
app_state.py, coordinates.py, constants.py, radio_config.py) - the
package's directory structure gave no visual signal about which
modules touch real hardware and which don't, even though hal/ already
existed as the conceptual boundary (protocols.py/fake.py/factory.py,
ARCHITECTURE.md §4.14, tests/hal/).

Moved all six into hal/ via git mv, so the package now physically
reflects that boundary: hal/ = Pi-specific, everything else under
radioglobe/ = pure app logic with zero hardware dependencies.

None of the six needed logic changes - Protocol's structural typing
means nothing had to subclass anything - only two relative imports
needed an extra `.` for the new nesting depth (positional_encoders.py:
.database -> ..database; display.py: .coordinates -> ..coordinates),
plus updating every importer's path (hal/factory.py's deferred
imports, main.py, and 9 test files) from `radioglobe.X` to
`radioglobe.hal.X`.

The existing "importing radioglobe.hal never triggers evdev/spidev/
liquidcrystal_i2c/vlc imports" guarantee still holds: hal/__init__.py
only imports .protocols and .factory, and factory.py's build_hardware()
still does its hardware-module imports inside the function body, not
at module scope - Python doesn't auto-import a package's submodules
just because the package itself was imported.

ARCHITECTURE.md updated throughout: repository layout tree, the
module dependency diagram (now shows hal/ as a subgraph), all six
per-module section headers, and the two paragraphs describing the
residual buttons.py-at-module-scope evdev import gap. README.md's
GPIO pin references updated too.
