+++ src/radioglobe/hal/dial.py
